Update to twilight 0.2 and modify CI to tag for organizations and tagged versions (#5)
This updates the dependencies and twilight-http to 0.2. For those willing to keep using the v6 API compatible version, a http-proxy:v6 tag has been pushed to https://hub.docker.com/r/twilightrs/http-proxy/tags with all dependencies updated except twilight-http.

It also modifies the CI config to allow for specifying a target on Dockerhub to push to that is not the username (e.g. the twilightrs org). DOCKERHUB_TOKEN was renamed to DOCKERHUB_PASSWORD as that is more accurate. It'll push the ARMv8 and AMD64 tag and then create a manifest called latest and one called v8. They are identical in content but as we move to future API versions, keeping the old v8 manifest and tagging v8-amd64 and v8-arm64 can be done.

Approved-by: Valdemar Erk <valdemar@erk.io>
Approved-by: Vivian Hellyer <vivian@hellyer.dev>
Merged-by: Jens Reidel <adrian@travitia.xyz>
Signed-off-by: Jens Reidel <adrian@travitia.xyz>